Bug 284537 NPE when double-clicking for diff in commit dialog
An optimization in JGit made this code try to look up disposed RevObjects
from the walker used by the file history. Instead of looking up HEAD
after receiving the full history we now look up the last change to the
relevant file. Further we only ask for the last change which makes this
much faster than it was before it was broken.

From: https://bugs.eclipse.org/284537
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>