Get rid of the class GitException
The class was based on the CVSException class from the package
org.eclipse.team.internal.ccvs.core, as well as SVNException
from org.tigris.subversion.subclipse.core.

Sadly the latter is not suitable as a basis when contributing
source code to the Eclipse Foundation, so it had to be removed.

If we need this class and its wrapper-helpers in the future we
can base it off CVSException. If we do, we should also add a
GitStatus, since TeamStatus is rally meant for API errors only.

From: https://bugs.eclipse.org/287743
Signed-off-by: Tor Arne Vestbø <torarnv@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>